% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{data_acceleration_all}
\alias{data_acceleration_all}
\title{Biometrical data of all volunteers.

A dataset containing the biometric data of x young adolescent volunteers
recorded in a social class-room situation.}
\format{A data frame containing "acceleration" data
for x, y, z, accelaration speed per time
\describe{
  \item{z'}{...}
  \item{z''}{...}
  ...
}}
\source{
\url{<url here>}
}
\usage{
data_acceleration_all
}
\description{
Biometrical data of all volunteers.

A dataset containing the biometric data of x young adolescent volunteers
recorded in a social class-room situation.
}
\references{
authors et al., 2019
}
\keyword{datasets}
