% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{data_instructions_all}
\alias{data_instructions_all}
\title{Biometrical data of all volunteers.

A dataset containing the biometric data of x young adolescent volunteers
recorded in a social class-room situation.}
\format{A data frame containing instructions that were given to the
volunteers and its timestamps
\describe{
  \item{""}{...}
    \item{}{}
  \item{z''}{...}
  ...
}}
\source{
\url{<url here>}
}
\usage{
data_instructions_all
}
\description{
Biometrical data of all volunteers.

A dataset containing the biometric data of x young adolescent volunteers
recorded in a social class-room situation.
}
\references{
authors et al., 2019
}
\keyword{datasets}
