% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edartefacts-package.R
\docType{package}
\name{edartefacts-package}
\alias{edartefacts}
\alias{edartefacts-package}
\title{edartefacts: Data Package Containing Biometry Data}
\description{
A datapackage containing biometry data.
}
\author{
\strong{Maintainer}: Marc Teunis \email{marc.teunis@hu.nl} (YOUR-ORCID-ID)

}
\keyword{internal}
