% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raw_data.R
\name{parse_raw_data}
\alias{parse_raw_data}
\title{Parse a zip archive containing the raw data per volunteer}
\usage{
parse_raw_data(path = NULL, path_out = NULL, volunteer_id = NULL)
}
\arguments{
\item{path}{A character directing to the path of the zip archive}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Parse a zip archive containing the raw data per volunteer
}
