% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{data_eda_all}
\alias{data_eda_all}
\title{Biometrical data of 1 volunteer.

A dataset containing the biometric data of x young adolescent volunteers
recorded in a social class-room situation.}
\format{A list with 2 data frames containing two elements "acceleration"
with acceleration data for x, y, z, accelaration speed per time and
"skin-conductance" with measured skin conductace from EDA by time
\describe{
  \item{""}{...}
    \item{}{}
  \item{z''}{...}
  ...
}}
\source{
\url{<url here>}
}
\usage{
data_eda_all
}
\description{
Biometrical data of 1 volunteer.

A dataset containing the biometric data of x young adolescent volunteers
recorded in a social class-room situation.
}
\references{
authors et al., 2019
}
\keyword{datasets}
