% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_raw_data.R
\name{parse_raw_data}
\alias{parse_raw_data}
\title{Parse a zip archive containing the raw data per volunteer}
\usage{
parse_raw_data(path = NULL, path_out = NULL, subject_id = NULL)
}
\arguments{
\item{path}{A character directing to zip archive containing data from
a single volunteer}

\item{path_out}{A character directing to the directory where the unzipped
files must be stored}

\item{volunteer_id}{A character single string indicating from which
volunteer you are parsing the data}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Parse a zip archive containing the raw data per volunteer
}
